+++ Source/Core/MultipartAPIRequest.swift
@@ -103,7 +103,7 @@ public class MultipartAPIRequest<Model: ResponseParseable, ErrorModel: ResponseP
                 failure?(apiError)
             } else if case .Success(let request, _, _) = completion {
                 let allPlugins = self.plugins + (self.tronDelegate?.plugins ?? [])
-                request.progress(progress).validate().handleResponse(success, failure: failure, responseBuilder: self.responseBuilder, errorBuilder: self.errorBuilder, plugins: allPlugins)
+                request.progress(progress).validate().handleResponse(success, failure: failure, dispatcher: self.dispatcher, responseBuilder: self.responseBuilder, errorBuilder: self.errorBuilder, plugins: allPlugins)
                 cancellableCallback(request)
             }
         }